Move the TrackMateConfigDialog to a common package.
